### 2023-05-09 #5457 @Kevin-Mattheus-Moerman
@hkabbech thanks for this submission. I am the AEiC for this track and here to help process initial steps. I have just triggered a scope review by the editorial board, to see if this work is deemed in scope. This is normal for submissions that are rather small e.g. in terms of lines of code. Over the coming ~2 weeks we'll review if this work conforms to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) .

In the meantime you can start fixing the potentially missing/invalid DOIs pointed out above :point_up: , you can call `@editorialbot check references` to check them again, and use `@editorialbot generate pdf` to update the paper.
