### 2023-05-10 #5457 @hkabbech
Thank you @Kevin-Mattheus-Moerman.

FYI this is a resubmission. I have added substantial analysis plots to the pipeline.

Regarding the missing DOI, I believe that there is no associated DOI to cite the keras library.

I am looking forward for your final decision.
