### 2023-05-30 #5457 @Kevin-Mattheus-Moerman
@hkabbech I have just removed the `query-scope` flag as the scope review concluded this work may be in scope for JOSS. However it does appear "borderline" in terms of what we typically accept (currently), and the reviewers and handling editor may still raise concerns on scope.
