+++ tracksegnet-env/pyvenv.cfg
@@ -1,8 +1,0 @@
-home = /usr/bin
-implementation = CPython
-version_info = 3.8.10.final.0
-virtualenv = 20.20.0
-include-system-site-packages = false
-base-prefix = /usr
-base-exec-prefix = /usr
-base-executable = /usr/bin/python3
