### 2023-12-05 cc390ac
close figures
