### 2023-12-05 0542701
debug function
