### 2023-12-10 2b8227a
ignore  Mean of empty slice warning
