### 2023-12-10 e2d63d1
Create dockerfile and docker-compose
