+++ CONTRIBUTING.md
@@ -1,0 +1,45 @@
+# Contributing Guidelines
+
+First, **thank you** for contributing!
+
+Before starting, you should read, agree to, and follow this guideline.
+
+---
+
+## How to contribute?
+
+### Report bugs
+
+Report bugs at: https://github.com/hkabbech/tracksegnet/issues/new.
+
+When reporting a bug, please include the detailed steps to reproduce the bug.
+Where possible, please write a test case.
+
+If you are not able to do that, that's fine! Open an issue anyway and let us
+know as much information as you can. We will get back to you to determine the
+problem, and (hopefully) fix it.
+
+
+### Submit feedback
+
+The best way to send feedback is to [create a new issue](https://github.com/hkabbech/tracksegnet/issues/new)
+on GitHub.
+
+If you are proposing a feature:
+
+* Explain how you see it working. Try to be as detailed as you can.
+* Try to keep the scope as narrow as possible. This will help make it easier to
+  implement.
+* Feel free to include any code you might already have, even if it is
+  just a rough idea. This is a volunteer-driven project, and contributions are
+  welcome!
+
+
+## Pull Request Guidelines
+
+Here are a few rules to follow in order to make code reviews and discussions go
+more smoothly before accepting and merging your work:
+
+* Include comments and documentation in your code.
+* Write commit messages that make sense, and rebase your branch before submitting your Pull Request.
+* While creating your Pull Request on GitHub, you must write a description which gives the context and/or explains why you are creating it.
