### 2023-12-10 2e296a0
add description + docker commands
