+++ requirements.txt
@@ -34,7 +34,7 @@ pyasn1-modules==0.2.8
 pyparsing==3.0.9
 python-dateutil==2.8.2
 pytz==2022.7.1
-requests==2.28.2
+requests==2.31.0
 requests-oauthlib==1.3.1
 rsa==4.9
 scipy==1.10.0
