+++ requirements.txt
@@ -1,7 +1,7 @@
 absl-py==1.4.0
 astunparse==1.6.3
 cachetools==5.3.0
-certifi==2022.12.7
+certifi>=2023.7.22
 charset-normalizer==3.0.1
 contourpy==1.0.7
 cycler==0.11.0
@@ -11,7 +11,7 @@ gast==0.4.0
 google-auth==2.16.0
 google-auth-oauthlib==0.4.6
 google-pasta==0.2.0
-grpcio==1.51.1
+grpcio>=1.53.0
 h5py==3.8.0
 idna==3.4
 importlib-metadata==6.0.0
@@ -27,14 +27,14 @@ oauthlib==3.2.2
 opt-einsum==3.3.0
 packaging==23.0
 pandas==1.5.3
-Pillow==9.4.0
+Pillow>=10.0.1
 protobuf==3.19.6
 pyasn1==0.4.8
 pyasn1-modules==0.2.8
 pyparsing==3.0.9
 python-dateutil==2.8.2
 pytz==2022.7.1
-requests==2.28.2
+requests>=2.31.0
 requests-oauthlib==1.3.1
 rsa==4.9
 scipy==1.10.0
@@ -43,13 +43,13 @@ six==1.16.0
 tensorboard==2.11.2
 tensorboard-data-server==0.6.1
 tensorboard-plugin-wit==1.8.1
-tensorflow==2.11.0
+tensorflow>=2.11.1
 tensorflow-estimator==2.11.0
 tensorflow-io-gcs-filesystem==0.30.0
 termcolor==2.2.0
 tqdm==4.64.1
 typing-extensions==4.4.0
-urllib3==1.26.14
-Werkzeug==2.2.3
+urllib3>=1.26.18
+werkzeug>=2.3.8
 wrapt==1.14.1
 zipp==3.11.0
