### 2023-12-10 #5457 @hkabbech
@Kevin-Mattheus-Moerman, many thanks for the updates!

> @hkabbech I am currently trying to find reviewers for this submission. If you would like to suggest any reviewers at this point, that would be hepful. Please mention their GitHub handles, but leave out the `@` symbol.

Unfortunately, I'm not sure if I know people who might help review my submission.

> In addition, please consider the below points:
>
>     * [x]  Please extend your README with a clear paragraph describing the software purpose/functionality.
>
>     * [x]  Please add contributing guidelines to your project. You can write a basic section in your README or you could link to a dedicated CONTRIBUTING.md file (see also: https://contributing.md/how-to-build-contributing-md/).

I modified the README accordingly and added a CONTRIBUTING.md file.
I also added a Dockerfile and docker-compose.yaml for the creation of a docker container running the code.
