### 2023-12-13 #5457 @ajasja
@Kevin-Mattheus-Moerman Yes, I could do it.
Does it matter that the method was already described in this article:
https://academic.oup.com/nar/article/51/20/10992/7288829

![image](https://github.com/openjournals/joss-reviews/assets/19926/0c254529-c6da-4af4-b7dc-162700046227)
