### 2023-12-13 #5457 @Kevin-Mattheus-Moerman
@ajasja thanks for volunteering to review. At first sight it looks like the core purpose of that newer paper is to describe the application of this work to a particular scientific research topic. I will review this in more detail. For the moment I'll ask the authors to point out the key difference/purpose for these papers.

@hkabbech could you please provide a brief but clear description of the major advances you've made to the software since this paper: doi.org/10.1038/s41598-019-53663-8 (I have this in the notes to the editor but would be good to reiterate this clearly here for the reviewers too), and could you clarify the key differences with the current paper and this new one: https://academic.oup.com/nar/article/51/20/10992/7288829 thanks.
