### 2023-12-15 #5457 @hkabbech
At the moment of submitting our work, the idea was to submit the software to JOSS for a better review of the tool/method, and a particular application (together with other experiments) to a biology journal where reviewers focus more on the discovery and biological methods employed. Both papers were submitted around the same time this year, while the biology paper got reviewed and published in only a few months (https://doi.org/10.1093/nar/gkad803), the pre-review process of the software paper was severely delayed.. Thus, in the meantime, we had to publish the code on Zenodo (https://zenodo.org/records/7767750).

The method was initially developed in our research group and published in 2019 (https://doi.org/10.1038/s41598-019-53663-8). The current software presents major improvements of our method and permits the replicability on other trajectory datasets. It contains also additional plotting measurements for interpretability of trajectory segmentation results.

We are still interested in getting our software reviewed by JOSS and hope this will still be possible.
