### 2023-12-18 #6157 @editorialbot
[REVIEW]: TrackSegNet: a tool for trajectory segmentation into diffusive states using supervised deep learning
**Submitting author:** @hkabbech (<a href="http://orcid.org/0000-0002-9200-2112">Hélène Kabbech</a>)
**Repository:** https://github.com/hkabbech/TrackSegNet
**Branch with paper.md** (empty if default branch):
**Version:** v1.3.0
