### 2024-02-08 hkabbech/TrackSegNet#7 @imagejan
Installation instructions for Windows
When running `docker compose build` and `docker compose run tracksegnet-env` on Windows, I get warnings like this:

```
2024-02-08 13:54:09.171592: W tensorflow/compiler/tf2tensorrt/utils/py_utils.cc:38] TF-TRT Warning: Cannot dlopen some TensorRT libraries. If you would like to use Nvidia GPU with TensorRT, please make sure the missing libraries mentioned above are installed properly.
```

It runs, but doesn't seem to utilize GPU.

Also, the alternative installation instructions mentioned in the README don't work on Windows (Command Prompt or PowerShell) as there's no `sudo` command available.

I tried installing with `pip install -r requirements.txt` in a conda environment, but again no GPU is used, in line with the warning on https://www.tensorflow.org/install/pip that versions 2.11 and later don't support GPU on Windows native.

I suggest the installation instructions could be refined to serve Windows users better. And even on Linux, there can be situations where users don't have `sudo` permissions, so maybe a conda-based installation procedure could be documented as well.
