### 2024-02-08 hkabbech/TrackSegNet#8 @imagejan
Add __pycache__ to .gitignore
I suggest adding the following line to `.gitignore`:

```
__pycache__
```

.. to avoid accidentally committing Python cache files to the git repository.
