### 2024-02-08 hkabbech/TrackSegNet#9 @imagejan
Proper Python packaging and deployment to PyPI/conda-forge
The project currently uses a `src` layout for most functionality, but doesn't use proper packaging. For example, the `tracksegnet.py` script currently does:

```
from src.utils import ...
```

.. instead of:

```
from utils import ...
```

.. which would be possible after `pip install -e .` if it was a proper package.

I suggest creating a `pyproject.toml` file and properly defining depenencies therein. (Tools like [`hatch`](https://hatch.pypa.io/latest/) can help setting things up according to best practices.)
