### 2024-02-08 hkabbech/TrackSegNet#10 @imagejan
Add tests and continous integration
Automated tests that run on every commit (e.g. using GitHub actions) greatly help maintaining software projects and avoiding future regressions. They're also a great way to include examples of using an application for others to look at.
