### 2024-02-08 hkabbech/TrackSegNet#11 @imagejan
Generate API documentation
For using the functionality of this repository in other projects, it is essential to have some API documentation accessible. Most of the methods have docstrings, so usage within an IDE is fine.
For better accessibility, I suggest also rendering a documentation website, using readthedocs and GitHub actions, for example.
