### 2024-02-08 #6157 @imagejan
I've opened these issues on the TrackSegNet repository:

* https://github.com/hkabbech/TrackSegNet/issues/7
* https://github.com/hkabbech/TrackSegNet/issues/8
* https://github.com/hkabbech/TrackSegNet/issues/9
* https://github.com/hkabbech/TrackSegNet/issues/10
* https://github.com/hkabbech/TrackSegNet/issues/11

My main concerns are:

* Installation instructions are incomplete for Windows (no GPU used with current installation procedure)
* There's no online documentation currently. For example, it would be helpful if the parameter description that is in the paper would also be accessible from the GitHub code repository. Also, the "statement of need" is only present in the paper.
* No (rendered) API documentation.
* No automated tests.
* No documented contribution guidelines.
* In the paper, the authors mention related software in the field, but do not describe in detail how their software compares to these other commonly-used packages.
