### 2024-02-16 ca6fe9c
Run toy example in git action #10
