### 2024-02-16 2c3954d
Define used ubuntu version
