### 2024-02-16 d2d582c
Setup readthedocs and gitaction on pull request #11
