+++ .github/workflows/.generate-readthedocs.yml
@@ -1,0 +1,16 @@
+name: generate_readthedocs
+on:
+  pull_request_target:
+    types:
+      - opened
+
+permissions:
+  pull-requests: write
+
+jobs:
+  pull-request-links:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: hkabbech/TrackSegNet/readthedocs-preview@v1
+        with:
+          project-slug: "tracksegnet"
