+++ .github/workflows/.generate-readthedocs.yml
@@ -3,6 +3,9 @@ on:
   pull_request_target:
     types:
       - opened
+      - reopened
+      - edited
+      - review_requested
 
 permissions:
   pull-requests: write
