### 2024-02-16 3ee555e
Remove line to install pylint #10
