+++ .github/workflows/.run-code.yml
@@ -24,7 +24,6 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
-        pip install pylint
     - name: Run code on toy example
       run: |
         python tracksegnet.py parms.csv
