+++ .github/workflows/.generate-readthedocs.yml
@@ -1,4 +1,5 @@
 name: generate_readthedocs
+
 on:
   pull_request_target:
     types:
@@ -12,7 +13,7 @@ permissions:
 
 jobs:
   pull-request-links:
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-22.04
     steps:
       - uses: hkabbech/TrackSegNet/readthedocs-preview@v1
         with:
