+++ docs/conf.py
@@ -1,22 +1,0 @@
-# -- General configuration ------------------------------------------------
-
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
-extensions = []
-
-project = 'Read the Docs'
-copyright = '2023, Read the Docs'
-author = 'H. Kabbech'
-
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-# These patterns also affect html_static_path and html_extra_path
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
-
-# -- Options for HTML output ----------------------------------------------
-
-# The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
-#
-html_theme = 'alabaster'
