+++ docs/source/generate_lstm_model.rst
@@ -1,0 +1,7 @@
+generate\_lstm\_model module
+============================
+
+.. automodule:: generate_lstm_model
+   :members:
+   :undoc-members:
+   :show-inheritance:
