+++ docs/source/experimental_tracks.rst
@@ -1,0 +1,7 @@
+experimental\_tracks module
+===========================
+
+.. automodule:: experimental_tracks
+   :members:
+   :undoc-members:
+   :show-inheritance:
