+++ docs/source/simulate_tracks.rst
@@ -1,0 +1,7 @@
+simulate\_tracks module
+=======================
+
+.. automodule:: simulate_tracks
+   :members:
+   :undoc-members:
+   :show-inheritance:
