+++ docs/source/index.rst
@@ -1,0 +1,30 @@
+.. tracksegnet documentation master file, created by
+   sphinx-quickstart on Fri Feb 16 23:09:32 2024.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+TrackSegNet's documentation!
+============================
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Readme:
+
+   readme
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Modules:
+
+   analysis
+   compute_features
+   experimental_tracks
+   generate_lstm_model
+   simulate_tracks
+   utils
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Contributing:
+
+   contributing
