+++ requirements.txt
@@ -53,3 +53,4 @@ urllib3>=1.26.18
 werkzeug>=2.3.8
 wrapt==1.14.1
 zipp==3.11.0
+myst_parser>=2.0.0
