+++ requirements.txt
@@ -54,3 +54,4 @@ werkzeug>=2.3.8
 wrapt==1.14.1
 zipp==3.11.0
 myst_parser>=2.0.0
+sphinx-rtd-theme
