### 2024-02-17 7fbab1d
Fix error empty data
