+++ docs/source/conf.py
@@ -19,7 +19,7 @@ release = '1.0.0'
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
-extensions = ["myst_parser"]
+extensions = ['myst_parser', 'sphinx.ext.autodoc']
 
 templates_path = ['_templates']
 exclude_patterns = []
