+++ src/analysis.py
@@ -1,6 +1,6 @@
-"""analysis functions
+"""Analysis and plotting functions
 
-This modules contains functions for trajectory analysis.
+This modules contains functions for trajectory analysis and plotting.
 """
 
 # Third-party modules
@@ -23,23 +23,41 @@ matplotlib.rcParams.update({
 })
 
 def convert_sigma(diffusion, parms):
-    """Convert diffusion to sigma value.
+    """Converts diffusion to sigma value.
 
-    :param diffusion: Diffusion D in um**2/s
+    :param diffusion: Diffusion value [um**2/s] to convert.
     :type diffusion: float
-    :return: sigma value
+    :param parms: Stored parameters containing global variables and instructions.
+    :type parms: dict
+    :return: sigma value.
     :rtype: float
     """
     sigma = np.sqrt((diffusion/parms['unit_diff'])*2)
     return sigma
 
 def convert_diffusion(sigma, parms):
-    """Convert sigma to diffusion value."""
+    """Converts sigma to diffusion value.
+
+    :param sigma: Sigma value to convert.
+    :type sigma: float
+    :param parms: Stored parameters containing global variables and instructions.
+    :type parms: dict
+    :return: Diffusion value [um**2/s].
+    :rtype: float
+    """
     diffusion = ((sigma**2)/2)*parms['unit_diff']
     return diffusion
 
 def compute_ptm(track_df, parms):
-    """Compute the probability transition matrix."""
+    """Computes the probability transition matrix (PTM).
+
+    :param track_df: Dataframe containing all extracted trajectories.
+    :type track_df: pd.DataFrame
+    :param parms: Stored parameters containing global variables and instructions.
+    :type parms: dict
+    :return: probabilities to transition between states
+    :rtype: pd.DataFrame
+    """
     print('\nCompute probability transition matrix (PTM)...')
     counts = np.zeros((parms['num_states'], parms['num_states']), dtype=int)
     for track_id in tqdm(track_df['track_id'].unique()):
@@ -61,8 +79,16 @@ def compute_ptm(track_df, parms):
     return ptm
 
 def make_tracklet_lists(track_df, parms):
-    """Segment the trajectories into tracklet (based on the state group) and regroup them within the
-    same list."""
+    """Segments the trajectories into tracklet (based on the state group) and regroup them within the
+    same list.
+
+    :param track_df: Dataframe containing all extracted trajectories.
+    :type track_df: pd.DataFrame
+    :param parms: Stored parameters containing global variables and instructions.
+    :type parms: dict
+    :return: List of tracklet groups. Each group consists of a dataframe of tracklets.
+    :rtype: list
+    """
     print('\nSegment and group tracklets per diffusive states...')
     tracklet_lists = [[] for _ in range(parms['num_states'])]
     for track_id in tqdm(track_df['track_id'].unique()):
@@ -88,8 +114,17 @@ def make_tracklet_lists(track_df, parms):
     return tracklet_lists
 
 def compute_msd(track, size, dim=2):
-    """Compute the mean square displacement (MSD) for a given track in order to estimate D and alpha
+    """Computes the mean square displacement (MSD) for a given track in order to estimate D and alpha
     using the formula:  log(MSD(dt)) ~ alpha.log(dt) + log(C), with C = 2nD.
+    
+    :param track: Dataframe containing a trajectory's coordinates.
+    :type track: pd.DataFrame
+    :param size: Number of delta time points to use for the MSD curve fit.
+    :type size: int
+    :param dim: Dimentionality of the track [Defaults: 2].
+    :type dim: int
[…]
