+++ docs/source/simulate_tracks.rst
@@ -1,5 +1,5 @@
-simulate\_tracks module
-=======================
+simulate\_tracks
+================
 
 .. automodule:: simulate_tracks
    :members:
