### 2024-02-26 984770e
Update installation instructions and fix tensorflow version #7
