### 2024-02-26 hkabbech/TrackSegNet#8 @hkabbech
Thanks for the reminder! The .gitignore file is now added to the repo.
I will close this issue.
