### 2024-02-26 hkabbech/TrackSegNet#11 @hkabbech
An API documentation containing the main purposes, instructions, module docstrings and the contribution guideline is now available at the following address: https://tracksegnet.readthedocs.io/en/latest/
For each pull request, an updated documentation is generated through a GitHub action
