### 2024-02-26 hkabbech/TrackSegNet#7 @hkabbech
The instructions have been updated for Windows, see https://tracksegnet.readthedocs.io/en/latest/installation.html
