+++ README.md
@@ -41,12 +41,11 @@ Requirement: python3.8 or equivalent and the virtualenv library
 
 ```bash
 # Create the environment:
-python -m venv tracksegnet-env
-# virtualenv -p /usr/bin/python3 tracksegnet-env
+python -m venv tracksegnet-env # virtualenv -p /usr/bin/python3 tracksegnet-env
 # Activate the environment:
 source ./tracksegnet-env/bin/activate # for Windows PowerShell: .\tracksegnet-env\Scripts\Activate.ps1 (run as administrator)
 # Install the required python libraries:
-python -m pip install -r requirements.txt
+pip install -e . # python -m pip install -r requirements.txt
 ```
 
 Note for later, to deactivate the virtual environment, type `deactivate`.
