### 2024-02-26 b725476
Add example plots to docs
