+++ docs/source/analysis-plots.md
@@ -1,0 +1,11 @@
+# Analysis plots – toy example
+
+![pipeline](img/fig_toy_example_displ.png)
+
+![pipeline](img/fig_toy_example_angles.png)
+
+![pipeline](img/fig_toy_example_vac.png)
+
+![pipeline](img/fig_toy_example_scatterplot_msd.png)
+
+![pipeline](img/fig_toy_example_tracklet_proportion.png)
