+++ Dockerfile
@@ -2,12 +2,7 @@ FROM python:3.8
 
 WORKDIR /env
 
-COPY requirements.txt ./
-RUN pip install -r requirements.txt
-
 COPY ./ ./
-# COPY data/ ./
-# COPY src/ ./
-# COPY tracksegnet.py ./
+RUN pip install -e .
 
-CMD ["python", "tracksegnet.py", "parms.csv"]
+CMD ["python", "tracksegnet-main.py", "parms.csv"]
