### 2024-03-01 hkabbech/TrackSegNet#9 @hkabbech
As suggested, a python package has now been created and published in pypi: https://pypi.org/project/tracksegnet/
