### 2024-03-01 #6157 @hkabbech
Dear @Kevin-Mattheus-Moerman,

I finished implementing the suggestions from @imagejan, the modifications include:
* An API documentation (regenerated at every pull request) containing the docstrings general information (introduction, instructions, parameter description as in the paper) and the contribution guidelines:
https://tracksegnet.readthedocs.io/en/latest/. (Issue 11)
* Updated Windows instructions in the README. (Issue 7)
* Creation and publication of a python package in pypi: https://pypi.org/project/tracksegnet/. (Issue 9)
* Test of the code quality by running pylint at every commit and test of the code executability on the toy example at every pull request. (Issue 10)

Regarding the related methods mentioned in the field, there are only a few at the moment and the methodologies are quite different from each other. The related software are not always user-friendly, making a fair comparison difficult to achieve, which was also not the initial purpose of this submission.
