### 2024-04-15 hkabbech/TrackSegNet#22 @ajasja
Virtual env activation script is in source ./tracksegnet-env/Scripts/activate on windows.
After installation of virtual env (on Windows isng python 3.11 the activation script is in source ./tracksegnet-env/Script/activate).

I would recommend changing the readme line to:
```
 source ./tracksegnet-env/bin/activate # Or source ./tracksegnet-env/Scripts/activate for Windows. For PowerShell: .\tracksegnet-env\Scripts\Activate.ps1 (run as administrator)
```
