### 2024-04-15 hkabbech/TrackSegNet#24 @ajasja
Installation with 3.10 works well.
