### 2024-04-15 #6157 @ajasja
@Kevin-Mattheus-Moerman @hkabbech
First, I would like to apologize for my slow response and commend the authors for having the patience to still publish this.

I have opened the following two minor issues dealing with installation problems:
https://github.com/hkabbech/TrackSegNet/issues/24
https://github.com/hkabbech/TrackSegNet/issues/22

My main concerns are:

*	A short description and links to other tools would be very welcome (no detailed benchmarks, just a paragraph about what the other tools do and how TrackSegNet differs).
*	In particular, the paper on line 29 the paper mentions major improvements to https://github.com/ismal/DL-MSS, but does not say what these improvements are.
In the companion paper (https://academic.oup.com/nar/article/51/20/10992/7288829) the authors go into more detail, which should also be mentioned in this paper:

>  which major changes include the computation of angles as an additional feature for better distinction of the trajectory confinement, the choice of using the MSD  analysis instead of the moment scaling spectrum (MSS) analysis, and the development of a user-friendly python software for replicability on other datasets.

Minor concerns:
*	It would be great if the raw tiff files were available in the toy examples (and not just the projection). I tried to open the tracks in MTrackJ, but was unable to do so as only one frame was available.
*	A brief description of the expected output (although it is pretty intuitive).
*	A mention of supported Python versions (for example, Python 3.11 does not currently work).

This may be outside the scope of the review, but I'm curious -- it seems that the parameters of the states need to be set in advance. Are these just initial values that are adjusted during the training? Probably not, as they seem to be used to create the synthetic trajectories. Is there a way to obtain these parameters from the experimental data? How would one obtain the parameters of the states for an unknown system?
