### 2024-04-16 #6157 @ajasja
Also, I forgot one more thing -- since the package is on pip now (which is great), it would be nice is this was mentioned in the installation instructions in the Readme.
