### 2024-05-09 5697547
Update command Windows script activation #22
