### 2024-05-09 d4f02db
Specify python versions #24 and precise installation with pip
