### 2024-05-09 #6157 @hkabbech
Dear all,

Apologies for this late reply, I had other priorities and couldn't find the time to finish this revision.

I have made some changes based on @ajsja comments.

- Fixed the open Issues regarding the installation
- Precised the supported python version in the README
- Added the installation instruction with pip

In `paper.md`:
- Precised the improvements regarding the initial first version of the method
- Improved the paragraph part on other tools/method from the literature
- Added a brief description of the expected output
- Fixed some grammatical mistakes

> It would be great if the raw tiff files were available in the toy examples (and not just the projection). I tried to open the tracks in MTrackJ, but was unable to do so as only one frame was available.

I am only in possession of the max. projection of the raw stack images which can be found in the data folder. The raw images would have been too large to upload in GitHub. To preview the trajectories in MTrackJ, you can create a hyperstack in imageJ (image>Hyperstacks>New Hyperstack...) with width=300px, height=50px, channels=1, slices=1 and frames=10,000. This could do the trick to visualize the tracks in time.

> This may be outside the scope of the review, but I'm curious -- it seems that the parameters of the states need to be set in advance. Are these just initial values that are adjusted during the training? Probably not, as they seem to be used to create the synthetic trajectories. Is there a way to obtain these parameters from the experimental data? How would one obtain the parameters of the states for an unknown system?

This is somewhat disadvantage of a supervised learning method, the experimental data must be generated based on the experimental data, however not much is known about their dynamics. During my PhD, I developed another method which is unsupervised (no need for synthetic data) based on the noise2noise denoising technique. The description of this other method can be found [here](https://pure.eur.nl/en/publications/identification-of-diffusive-states-in-tracking-applications-using), however this method is not yet available as a software tool since it requires more developments and tests from our side.

Thank you for all your inputs!

If there is no more changes to be made, I will update the documentation and the pypi link with a new version.
