### 2024-05-30 #6157 @Kevin-Mattheus-Moerman
@hkabbech can you work on your points above, and also check that potential DOI issue :point_up:
