### 2024-06-03 e775de0
Remove affiliation number
