+++ docs/source/conf.py
@@ -15,7 +15,7 @@ sys.path.insert(0, os.path.abspath('../../src/tracksegnet'))
 project = 'tracksegnet'
 copyright = '2024, H. Kabbech'
 author = 'H. Kabbech'
-release = '1.0.0'
+release = '1.3.0'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
