+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "tracksegnet"
-version = "1.2.5"
+version = "1.3.0"
 dynamic = ["dependencies"]
 authors = [
   { name="Hélène Kabbech", email="helene.kabbech@gmail.com" },
