### 2024-06-03 #6157 @hkabbech
Done!
The latest version is 1.3.0, and the DOI is [10.5281/zenodo.7767749](https://zenodo.org/doi/10.5281/zenodo.7767749).
