### 2024-06-04 #6157 @Kevin-Mattheus-Moerman
@hkabbech looks like this PR might fix the paper compilation error: https://github.com/hkabbech/TrackSegNet/pull/27
