### 2024-06-04 #6157 @Kevin-Mattheus-Moerman
@hkabbech  I have checked your repository, the archive link, and the paper. Most seems in order, however the below point requires your attention:

- [ ] In the paper, check `confinment` which should perhaps be `confinement`
